[CHORE] mypage completed
